chore: reorganize flask_api for dockerization
