+++ flask_api/pyproject.toml
@@ -13,6 +13,17 @@ flask = "^2.3.2"
 pandas = "^2.0.3"
 xgboost = "^1.7.6"
 
+[tool.poetry.group.dev.dependencies]
+mypy = "^1.4.1"
+black = "^23.3.0"
+pylint = "^2.17.4"
+pytest = "^7.4.0"
+flake8 = "^6.0.0"
+types-openpyxl = "^3.1.0.13"
+pandas-stubs = "^2.0.2.230605"
+types-pygments = "^2.15.0.1"
+types-colorama = "^0.4.15.11"
+types-setuptools = "^68.0.0.0"
 
 [build-system]
 requires = ["poetry-core"]
