--- flask_api/utils/__init__.py
@@ -1,5 +1,0 @@
-# flake8: noqa
-"""
-   gestion des imports modules 
-"""
-from .controllers import home, predire_credit
